Changes: - services.rs -state.rs -compilations.rs -rapport.rs
services.rs
Chat has now a map for keywords and is able to provide some limited data about mitigations from the user input.

Prompting has been changed for how it interacts with the LLM and the user.

State.rs
Changed so that the Chat is able to access data from the state, where the data from the file is stored.

Compilation.rs
Changed so that the Chat works with the app state.

rapport.rs
Removed some parameters, citation and likelyhood, since we don't need to provice them, instead we provide the top 10.